Created pkg directory to store common functions and structs that can be used by both iDRAC-Reference-Tools and IRC-Reference-Tools. Ideally pkg and internal would be moved to a separate repo. But for now, IRC imports iDRAC-Reference-Tools and accesses the pks directory
