+++ cmd/dbdiscauth/dbdiscauth.go
@@ -5,7 +5,6 @@ package main
 
 import (
 	"database/sql"
-	"encoding/json"
 	"fmt"
 	"log"
 	"os"
@@ -17,6 +16,7 @@ import (
 
 	"github.com/dell/iDRAC-Telemetry-Reference-Tools/internal/auth"
 	"github.com/dell/iDRAC-Telemetry-Reference-Tools/internal/messagebus/stomp"
+	"github.com/dell/iDRAC-Telemetry-Reference-Tools/pkg/dbdiscauth"
 )
 
 var configStrings = map[string]string{
@@ -49,60 +49,60 @@ func getHECInstancesFromDB(db *sql.DB) ([]auth.SplunkConfig, error) {
 	return ret, nil
 }
 
-func getInstancesFromDB(db *sql.DB) ([]auth.Service, error) {
-	results, err := db.Query("SELECT serviceType, ip, authType, auth FROM services")
-	if err != nil {
-		return nil, err
-	}
+// func getInstancesFromDB(db *sql.DB) ([]auth.Service, error) {
+// 	results, err := db.Query("SELECT serviceType, ip, authType, auth FROM services")
+// 	if err != nil {
+// 		return nil, err
+// 	}
 
-	var ret []auth.Service
-	for results.Next() {
-		var value auth.Service
-		var tmp string
-		err = results.Scan(&value.ServiceType, &value.Ip, &value.AuthType, &tmp)
-		if err != nil {
-			return nil, err
-		}
-		err := json.Unmarshal([]byte(tmp), &value.Auth)
-		if err != nil {
-			return nil, err
-		}
-		ret = append(ret, value)
-	}
-	return ret, nil
-}
+// 	var ret []auth.Service
+// 	for results.Next() {
+// 		var value auth.Service
+// 		var tmp string
+// 		err = results.Scan(&value.ServiceType, &value.Ip, &value.AuthType, &tmp)
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		err := json.Unmarshal([]byte(tmp), &value.Auth)
+// 		if err != nil {
+// 			return nil, err
+// 		}
+// 		ret = append(ret, value)
+// 	}
+// 	return ret, nil
+// }
 
-func deleteServiceFromDB(db *sql.DB, service auth.Service, authService *auth.AuthorizationService) error {
-	stmt, err := db.Prepare("DELETE FROM services WHERE ip = ?")
-	if err != nil {
-		return err
-	}
-	_, err = stmt.Exec(service.Ip)
-	if err != nil {
-		return err
-	}
-	return nil
-}
+// func deleteServiceFromDB(db *sql.DB, service auth.Service, authService *auth.AuthorizationService) error {
+// 	stmt, err := db.Prepare("DELETE FROM services WHERE ip = ?")
+// 	if err != nil {
+// 		return err
+// 	}
+// 	_, err = stmt.Exec(service.Ip)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	return nil
+// }
 
-func addServiceToDB(db *sql.DB, service auth.Service, authService *auth.AuthorizationService) error {
-	stmt, err := db.Prepare("INSERT INTO services(serviceType, ip, authType, auth) VALUES(?, ?, ?, ?)")
-	if err != nil {
-		return err
-	}
-	jsonStr, err := json.Marshal(service.Auth)
-	if err != nil {
-		return err
-	}
-	_, err = stmt.Exec(service.ServiceType, service.Ip, service.AuthType, string(jsonStr))
-	if err != nil {
-		return err
-	}
-	_ = authService.SendService(service)
-	return nil
-}
+// func addServiceToDB(db *sql.DB, service auth.Service, authService *auth.AuthorizationService) error {
+// 	stmt, err := db.Prepare("INSERT INTO services(serviceType, ip, authType, auth) VALUES(?, ?, ?, ?)")
+// 	if err != nil {
+// 		return err
+// 	}
+// 	jsonStr, err := json.Marshal(service.Auth)
+// 	if err != nil {
+// 		return err
+// 	}
+// 	_, err = stmt.Exec(service.ServiceType, service.Ip, service.AuthType, string(jsonStr))
+// 	if err != nil {
+// 		return err
+// 	}
+// 	_ = authService.SendService(service)
+// 	return nil
+// }
 
 // splunk configuration are getting added in the database
-//TO DO: Update it in the db
+// TO DO: Update it in the db
 func splunkAddHECToDB(db *sql.DB, SplunkConfig auth.SplunkConfig, authService *auth.AuthorizationService) error {
 	stmt, err := db.Prepare("INSERT INTO HttpEventCollector(url, `key`, `index`) VALUES(?, ?, ?)")
 	if err != nil {
@@ -235,7 +235,7 @@ func main() {
 	}
 
 	//Fetch and publish configured services in the database
-	authServices, err := getInstancesFromDB(db)
+	authServices, err := dbdiscauth.GetInstancesFromDB(db)
 	if err != nil {
 		log.Print("Failed to get db entries: ", err)
 	} else {
@@ -252,7 +252,7 @@ func main() {
 		log.Printf("Received command in dbdiscauth: %s", command.Command)
 		switch command.Command {
 		case auth.RESEND:
-			authServices, err := getInstancesFromDB(db)
+			authServices, err := dbdiscauth.GetInstancesFromDB(db)
 			if err != nil {
 				log.Print("Failed to get db entries: ", err)
 				break
@@ -261,12 +261,12 @@ func main() {
 				go authorizationService.SendService(element) //nolint: errcheck
 			}
 		case auth.ADDSERVICE:
-			err = addServiceToDB(db, command.Service, authorizationService)
+			err = dbdiscauth.AddServiceToDB(db, command.Service, authorizationService)
 			if err != nil {
 				log.Print("Addservice,Failed to write db entries: ", err)
 			}
 		case auth.DELETESERVICE:
-			err = deleteServiceFromDB(db, command.Service, authorizationService)
+			err = dbdiscauth.DeleteServiceFromDB(db, command.Service, authorizationService)
 			if err != nil {
 				log.Print("Deleteservice Failed to delete db entries: ", err)
 			}
